# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AdCraft AI - Video Generation API
  description: |
    AI-powered video generation platform that transforms product images into professional commercials.
    
    ## Features
    - **AI Video Generation**: Create professional commercials from product images using Google Veo AI
    - **Interactive Chat**: Refine video concepts through natural language conversation
    - **Real-time Status**: Track video generation progress with WebSocket-like polling
    - **Multi-language Support**: Full English and Japanese localization
    - **Cost Management**: Built-in budget tracking and cost optimization
    - **Production Monitoring**: Comprehensive health checks and performance metrics
    
    ## Authentication
    Most endpoints are publicly accessible for the demo application.
    Admin endpoints require API key authentication via Bearer token.
    
    ## Rate Limits
    - Video Generation: 1 request per hour per IP, 3 per day maximum
    - Chat Refinement: 10 messages per session maximum
    - Status Checks: 1 request per second per job
    - Admin Endpoints: Requires valid admin API key
    
    ## Cost Information
    - Average cost per video generation: $1.81-2.01
    - Budget alerts at 50%, 75%, 90% thresholds
    - Real-time cost tracking available through admin dashboard
  version: 1.0.0
  contact:
    name: AdCraft AI Support
    url: https://github.com/your-org/adcraft-ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
paths:
...
  /admin/monitoring:
    get:
      tags:
        - Administration
      summary: Get Monitoring Dashboard Data
      description: |
        Retrieves comprehensive monitoring data for system administration.
        Requires admin authentication via Bearer token.
      operationId: getMonitoringDashboard
      security:
        - BearerAuth: []
      parameters:
        - name: section
          in: query
          required: false
          description: Specific monitoring section to retrieve
          schema:
            type: string
            enum: [overview, health, performance, costs, security, alerts, logs]
            example: "overview"
        - name: timeRange
          in: query
          required: false
          description: Time range for historical data
          schema:
            type: string
            default: "24h"
            example: "24h"
        - name: details
          in: query
          required: false
          description: Include detailed historical data
          schema:
            type: boolean
            default: false
            example: false
      responses:
        '200':
          description: Monitoring data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringDashboard'
        '401':
          description: Unauthorized - invalid or missing admin API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid section parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Administration
      summary: Execute Monitoring Actions
      description: |
        Executes administrative monitoring actions such as force health checks,
        clearing trends, updating configuration, or resolving alerts.
      operationId: executeMonitoringAction
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonitoringActionRequest'
            example:
              action: "force_health_check"
      responses:
        '200':
          description: Action executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringActionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid action or parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin API key for accessing administrative endpoints
  schemas:
...
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy, critical]
          example: "healthy"
        timestamp:
          type: string
          format: date-time
          example: "2025-08-28T10:30:00.000Z"
        uptime:
          type: integer
          description: Uptime in seconds
          example: 86400
        overallScore:
          type: integer
          minimum: 0
          maximum: 100
          description: Overall health score
          example: 95
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceHealth'
    ServiceHealth:
      type: object
      properties:
        name:
          type: string
          example: "Vertex AI"
        status:
          type: string
          enum: [healthy, degraded, unhealthy, critical]
          example: "healthy"
        responseTime:
          type: integer
          description: Response time in milliseconds
          example: 245
        lastCheck:
          type: string
          format: date-time
          example: "2025-08-28T10:30:00.000Z"
        details:
          type: object
          properties:
            available:
              type: boolean
              example: true
            errorRate:
              type: number
              format: float
              example: 0.02

    # Admin Dashboard Schemas
    MonitoringDashboard:
      type: object
      properties:
        system:
          $ref: '#/components/schemas/SystemStatus'
        performance:
          $ref: '#/components/schemas/PerformanceMetrics'
        costs:
          $ref: '#/components/schemas/CostMetrics'
        security:
          $ref: '#/components/schemas/SecurityMetrics'
        alerts:
          $ref: '#/components/schemas/AlertSummary'
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
    SystemStatus:
      type: object
      properties:
        nodeVersion:
          type: string
          example: "v18.17.0"
        platform:
          type: string
          example: "linux"
        uptime:
          type: number
          example: 86400
        memoryUsage:
          type: object
          properties:
            rss:
              type: integer
              example: 52428800
            heapTotal:
              type: integer
              example: 41943040
            heapUsed:
              type: integer
              example: 29855488
    PerformanceMetrics:
      type: object
      properties:
        totalRequests:
          type: integer
          example: 1247
        averageResponseTime:
          type: number
          format: float
          example: 234.5
        errorRate:
          type: number
          format: float
          example: 2.3
        currentCpuUsage:
          type: number
          format: float
          example: 45.2
        currentMemoryUsage:
          type: integer
          example: 512000000
    CostMetrics:
      type: object
      properties:
        totalBudget:
          type: number
          format: float
          example: 300.00
        currentSpend:
          type: number
          format: float
          example: 45.50
        remainingBudget:
          type: number
          format: float
          example: 254.50
        percentageUsed:
          type: number
          format: float
          example: 15.17
        alertLevel:
          type: string
          enum: [safe, warning, critical]
          example: "safe"
    SecurityMetrics:
      type: object
      properties:
        totalEvents:
          type: integer
          example: 156
        eventsBySeverity:
          type: object
          properties:
            low:
              type: integer
              example: 98
            medium:
              type: integer
              example: 42
            high:
              type: integer
              example: 14
            critical:
              type: integer
              example: 2
        eventsLast24h:
          type: integer
          example: 23
        eventsLastHour:
          type: integer
          example: 3
    AlertSummary:
      type: object
      properties:
        totalActive:
          type: integer
          example: 3
        bySeverity:
          type: object
          properties:
            low:
              type: integer
              example: 1
            medium:
              type: integer
              example: 1
            high:
              type: integer
              example: 1
            critical:
              type: integer
              example: 0
        recent:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
    Alert:
      type: object
      properties:
        id:
          type: string
          example: "alert_abc123"
        ruleId:
          type: string
          example: "high-error-rate"
        title:
          type: string
          example: "High Error Rate Detected"
        message:
          type: string
          example: "Error rate has exceeded 5% threshold"
        severity:
          type: string
          enum: [low, medium, high, critical]
          example: "high"
        status:
          type: string
          enum: [active, resolved]
          example: "active"
        triggeredAt:
          type: string
          format: date-time
          example: "2025-08-28T10:25:00.000Z"
...
    MonitoringActionRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [force_health_check, clear_trends, update_config, resolve_alert, export_data]
          example: "force_health_check"
        config:
          type: object
          description: Configuration for update_config action
        alertId:
          type: string
          description: Alert ID for resolve_alert action
        source:
          type: string
          description: Alert source for resolve_alert action
        startDate:
          type: string
          format: date-time
          description: Start date for export_data action
        endDate:
          type: string
          format: date-time
          description: End date for export_data action
    MonitoringActionResponse:
      type: object
      properties:
        message:
          type: string
          example: "Health check completed"
        healthCheck:
          $ref: '#/components/schemas/HealthResponse'
        resolved:
          type: boolean
          example: true
        data:
          type: object
          description: Exported data for export_data action
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
...
    ResponseMetadata:
      type: object
      properties:
        correlationId:
          type: string
          example: "corr_abc123"
        timestamp:
          type: string
          format: date-time
          example: "2025-08-28T10:30:00.000Z"
        generatedIn:
          type: integer
          description: Response generation time in milliseconds
          example: 145
        version:
          type: string
          example: "1.0.0"

    # Error Response Schemas
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Product name is required and must be between 1-100 characters"
        correlationId:
          type: string
          example: "corr_abc123"
        details:
          type: object
          description: Additional error details (development only)
...